+++ aif360/algorithms/inprocessing/gerryfair_classifier.py
@@ -1,5 +1,4 @@
 import numpy as np
-import pandas as pd
 from sklearn import linear_model
 import aif360.algorithms.inprocessing.gerryfair.clean as clean
 import aif360.algorithms.inprocessing.gerryfair.heatmap as heatmap
@@ -8,11 +7,12 @@ from aif360.algorithms.inprocessing.gerryfair.auditor import Auditor
 from aif360.algorithms.inprocessing.gerryfair.classifier_history import ClassifierHistory
 from aif360.algorithms import Transformer
 import matplotlib
-
 try:
     matplotlib.use('TkAgg')
 except:
     print("Matplotlib Error, comment out matplotlib.use('TkAgg')")
+
+
 
 
 class Model(Transformer):
